온라인 엔티티 제작중

+++ src/main/java/com/gcf/spring/repository/BannerTwoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.gcf.spring.entity.BannerTwo;
 
-public interface BannerTwoRepository extends JpaRepository<BannerTwo, Long> {
+public interface BannerTwoRepository extends JpaRepository<BannerTwo, Integer> {
 
 }
